COLORADO-1408 Eposed ability to add a sub collection and updated the form
The islandora_coalliance_manager module functionality to add a collection to
a parent collection was expanded to allow the passing of the parent collection
to the drupal form.  This allowed the setting of the desired parent collection
and desired namespace selection boxes in the form to refer to the parent
collection and the parent collection's namespace.

There were some code changes unrelated to this feature that were incorporated
to reflect the code that was currently running on the Colorado server but was
not under source control.  I do not know where they came from but they were
current.